+++ app/scripts/dimApp.i18n.js
@@ -96,6 +96,7 @@
             BackToDIM: "Back to DIM",
             CannotMove: "Cannot move that item off this character.",
             Drag: "Hold shift or pause over drop zone to transfer a partial stack.",
+            FilterCount: "{{amount}} matching current filter",
             ChangingPerks: "Changing Perks Not Supported",
             ChangingPerksInfo: "Sorry, there's no way to change perks outside the game. We wish we could!",
             NeverShow: "Never show me this again.",