+++ app/scripts/store/dimStores.directive.js
@@ -21,6 +21,8 @@
         '  <div ng-repeat="(category, buckets) in ::vm.buckets.byCategory track by category" class="section" ng-class="::category | lowercase">',
         '    <div class="title">',
         '      <span class="collapse-handle" ng-click="vm.toggleSection(category)"><i class="fa collapse" ng-class="vm.settings.collapsedSections[category] ? \'fa-plus-square-o\': \'fa-minus-square-o\'"></i> <span translate="Bucket.{{::category}}"></span></span>',
+        //     if the user has input a search query, display the amount of items that match the query
+        '      <span ng-hide="vm.search.query === \'\'" class="filter-count">(<span translate="Help.FilterCount" translate-values="{ amount: vm.search.filterCounts[category] }"></span>)</span>',
         '      <span ng-if="::vm.vault.vaultCounts[category] !== undefined" class="bucket-count">{{ vm.vault.vaultCounts[category] }}/{{::vm.vault.capacityForItem({sort: category})}}</span>',
         '    </div>',
         '    <div class="store-row items" ng-if="!vm.settings.collapsedSections[category]" ng-repeat="bucket in ::buckets track by bucket.id" ng-repeat="bucket in ::buckets track by bucket.id"><i ng-click="vm.toggleSection(bucket.id)" class="fa collapse" ng-class="vm.settings.collapsedSections[bucket.id] ? \'fa-plus-square-o\': \'fa-minus-square-o\'"></i>',
@@ -55,9 +57,9 @@
     }
   }
 
-  StoresCtrl.$inject = ['dimSettingsService', '$scope', 'dimStoreService', 'dimPlatformService', 'loadingTracker', 'dimBucketService', 'dimInfoService', '$translate'];
+  StoresCtrl.$inject = ['dimSettingsService', '$scope', 'dimStoreService', 'dimPlatformService', 'loadingTracker', 'dimBucketService', 'dimInfoService', 'dimSearchService', '$translate'];
 
-  function StoresCtrl(settings, $scope, dimStoreService, dimPlatformService, loadingTracker, dimBucketService, dimInfoService, $translate) {
+  function StoresCtrl(settings, $scope, dimStoreService, dimPlatformService, loadingTracker, dimBucketService, dimInfoService, dimSearchService, $translate) {
     var vm = this;
     const didYouKnowTemplate = `<p>${$translate.instant('DidYouKnow.Collapse')}</p>` +
                                `<p>${$translate.instant('DidYouKnow.Expand')}</p>`;
@@ -77,6 +79,7 @@
     dimBucketService.then(function(buckets) {
       vm.buckets = angular.copy(buckets);
     });
+    vm.search = dimSearchService;
     vm.toggleSection = function(id) {
       didYouKnow();
       vm.settings.collapsedSections[id] = !vm.settings.collapsedSections[id];